+++ web/src/main/kotlin/sashastory/dev/model/Form.kt
@@ -3,6 +3,7 @@ package sashastory.dev.model
 import com.github.vok.framework.sql2o.Entity
 import com.github.vok.framework.sql2o.Table
 import java.io.Serializable
+import java.math.BigDecimal
 import javax.validation.constraints.Max
 
 /**
@@ -12,7 +13,7 @@ import javax.validation.constraints.Max
 
 @Table("FORM")
 data class Form(
-        override var id: Long? = null,
+        override var id: BigDecimal? = null,
 
         var formNumber: String? = null,
 
@@ -23,4 +24,4 @@ data class Form(
 
         var schoolId: Long? = null
 
-) : Entity<Long>, Serializable
+) : Entity<BigDecimal>, Serializable
